popUpFunction.js updated
