+++ src/index.js
@@ -1,4 +1,6 @@
 import './index.css';
 import getPokemon from './modules/getPokemon';
+// import formFunction from './modules/addComment';
 
 window.addEventListener('DOMContentLoaded', getPokemon);
+// formFunction();
